Instance

Create a Vulkan instance and enumerate the instance's supported extensions.

Module Reference

VulkanExperiments.DrawingATriangle.Instance — Module

Instance

This module is the second step for getting started with Vulkan. It's purpose is to create a window with a Vulkan instance.

See https://vulkan-tutorial.com/en/Drawingatriangle/Setup/Instance

VulkanExperiments.DrawingATriangle.Instance.App — Type

The application's state. Anything the main app needs to run will be included here.

VulkanExperiments.DrawingATriangle.Instance.buildApp — Method

Construct a new application instance.

VulkanExperiments.DrawingATriangle.Instance.cleanup — Method

cleanup(app::App) :: App

Cleanup any and all graphical resources.

VulkanExperiments.DrawingATriangle.Instance.initVulkan — Method

initVulkan(app::App) :: App

Initialize Vulkan resources for this application.

VulkanExperiments.DrawingATriangle.Instance.initWindow — Method

initWindow(app::App) :: App

Create the application's main window. Configure to support Vulkan instead of OpenGL.

VulkanExperiments.DrawingATriangle.Instance.main — Method

main()

The application's entry point.

VulkanExperiments.DrawingATriangle.Instance.mainLoop — Method

mainLoop(app::App) :: App

The application's main loop. This function blocks until the window is closed by clicking the 'x' or similarly exited.